6718 regroup authorize test
